Add files via upload
Used C++ to read data from several .txt files and make a camera at the surface of the earth and let it take pictures of the whole universe, stars and moon will be in the picture. Then I let camera fly to the moon, as a result, the moon became larger and larger in the picture. I made about 1000 .pgm files during the process of the camera flew toward the moon, then I incorporated these .pgm files to make a 30-second film.